新增第三方登录账号表

+++ database/migrations/2019_10_02_200001_alter_users_tables.php
@@ -5,6 +5,7 @@ use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 use Z1px\App\Models\Users\UsersModel;
 use Z1px\App\Models\Users\UsersPassportsModel;
+use Z1px\App\Models\Users\UsersThirdModel;
 
 
 class AlterUsersTables extends Migration
@@ -12,6 +13,7 @@ class AlterUsersTables extends Migration
 
     private $users_model = UsersModel::class;
     private $users_passports_model = UsersPassportsModel::class;
+    private $users_third_model = UsersThirdModel::class;
 
     /**
      * Run the migrations.
@@ -24,12 +26,17 @@ class AlterUsersTables extends Migration
          * 更新用户表
          */
         if(Schema::hasTable(app($this->users_passports_model)->getTable())
-            && Schema::hasTable(app($this->users_model)->getTable())) {
+            && Schema::hasTable(app($this->users_model)->getTable())
+            && Schema::hasTable(app($this->users_third_model)->getTable())) {
             Schema::table(app($this->users_passports_model)->getTable(), function (Blueprint $table) {
                 # 创建外键约束
                 $table->foreign('user_id')
                     ->references('id')
                     ->on(app($this->users_model)->getTable())
+                    ->onDelete('cascade');
+                $table->foreign('user_id')
+                    ->references('id')
+                    ->on(app($this->users_third_model)->getTable())
                     ->onDelete('cascade');
             });
         }
